requirements.txt
gradio
transformers
datasets
peft
pandas
openpyxl
accelerate
bitsandbytes
Pillow
qwen-vl-utils